Fix bug in doPopup where I was ignoring the dpQueryState that was passed in.
NOTE: Facets with selections, when edited, need their dpQueryState to be stripped of the criteria for the facet being edited, while facets without selections need to have the entire current state in them. doPopup() was being called with the proper data, but I had started dropping it on the floor in 01ed4f5.